Extract CommandPipeline from controller.py (#394 Phase 2E)
New CommandPipeline class owns the command queue, active processes,
move processes, and related state. Controller.process() becomes:

    self.__pipeline.propagate_exceptions()
    self.__pipeline.cleanup()
    self.__pipeline.step()
    self.__update_model()

Methods moved (10):
- queue, step (was __process_commands), cleanup (was __cleanup_commands)
- propagate_exceptions, spawn_deferred_move, _spawn_move_process
- _get_pair_context_for_command, _get_pair_context_for_file
- _find_pair_by_id, _build_extract_request

State moved (5):
- command_queue, active_command_processes, active_move_processes
- moved_file_keys, pending_validation_keys

Controller retains Command, CommandProcessWrapper, and
_MAX_CONCURRENT_COMMAND_PROCESSES (external callers reference them).
_sync_persist_to_all_builders passed as callback.

controller.py: 1125 → 764 lines (-361)
New file: controller/command_pipeline.py (450 lines)
297 tests pass.
